refactor: cart methods moved to a separate file
